feat(folder structure): moves auth files into src folder
