+++ index.tsx
@@ -1,7 +1,7 @@
-import SocialAuth from './socialLogin'
-import InstagramAuth from './instagram'
-import GoogleAuth from './google'
-import FacebookAuth from './facebook'
-import AppleAuth from './apple'
+import AppleAuth from './src/apple'
+import FacebookAuth from './src/facebook'
+import GoogleAuth from './src/google'
+import InstagramAuth from './src/instagram'
+import SocialAuth from './src/socialLogin'
 
 export { SocialAuth, InstagramAuth, GoogleAuth, FacebookAuth, AppleAuth }
